+++ cumulusci/cli/tests/test_task.py
@@ -1,14 +1,14 @@
-"""Tests for the RunTaskCommand class"""
-
-from cumulusci.cli.runtime import CliRuntime
-from cumulusci.cli.cci import RunTaskCommand
-import click
-import pytest
 from unittest.mock import Mock, patch
 
-from cumulusci.cli import cci
+import click
+import pytest
+
+from cumulusci.core.config import BaseProjectConfig
 from cumulusci.core.exceptions import CumulusCIUsageError
-from cumulusci.cli.tests.utils import DummyTask
+from cumulusci.cli.runtime import CliRuntime
+
+from .. import task
+from .utils import DummyTask, run_click_command
 
 color_opts = {"options": {"color": {}}}
 multiple_opts = {"options": {"foo": {}, "bar": {}, "baz": {}}}
@@ -31,13 +31,14 @@ def runtime():
 
     runtime.keychain = Mock()
     runtime.keychain.get_default_org.return_value = (None, None)
+    runtime._load_keychain = Mock()
 
     yield runtime
 
 
 def test_task_run(runtime):
     DummyTask._run_task = Mock()
-    multi_cmd = cci.RunTaskCommand()
+    multi_cmd = task.RunTaskCommand()
     with click.Context(multi_cmd, obj=runtime) as ctx:
         cmd = multi_cmd.get_command(ctx, "dummy-task")
         cmd.callback(runtime, "dummy-task", color="blue")
@@ -48,7 +49,7 @@ def test_task_run(runtime):
 def test_task_run__no_project(runtime):
     runtime.project_config = None
     runtime.project_config_error = Exception("Broken")
-    multi_cmd = cci.RunTaskCommand()
+    multi_cmd = task.RunTaskCommand()
     with pytest.raises(Exception, match="Broken"):
         with click.Context(multi_cmd, obj=runtime) as ctx:
             multi_cmd.get_command(ctx, "dummy-task")
@@ -56,7 +57,7 @@ def test_task_run__no_project(runtime):
 
 def test_task_run__debug_before(runtime):
     DummyTask._run_task = Mock()
-    multi_cmd = cci.RunTaskCommand()
+    multi_cmd = task.RunTaskCommand()
     set_trace = Mock(side_effect=SetTrace)
 
     with click.Context(multi_cmd, obj=runtime) as ctx:
@@ -74,7 +75,7 @@ def test_task_run__debug_before(runtime):
 
 def test_task_run__debug_after(runtime):
     DummyTask._run_task = Mock()
-    multi_cmd = cci.RunTaskCommand()
+    multi_cmd = task.RunTaskCommand()
     set_trace = Mock(side_effect=SetTrace)
 
     with click.Context(multi_cmd, obj=runtime) as ctx:
@@ -92,7 +93,7 @@ def test_task_run__debug_after(runtime):
 
 def test_task_run__help(runtime):
     DummyTask._run_task = Mock()
-    multi_cmd = cci.RunTaskCommand()
+    multi_cmd = task.RunTaskCommand()
     with click.Context(multi_cmd, obj=runtime) as ctx:
         cmd = multi_cmd.get_command(ctx, "dummy-task")
 
@@ -100,16 +101,16 @@ def test_task_run__help(runtime):
 
 
 def test_task_run__list_commands(runtime):
-    multi_cmd = cci.RunTaskCommand()
+    multi_cmd = task.RunTaskCommand()
     with click.Context(multi_cmd, obj=runtime) as ctx:
         commands = multi_cmd.list_commands(ctx)
     assert commands == ["dummy-derived-task", "dummy-task"]
 
 
 def test_format_help(runtime):
-    with patch("cumulusci.cli.cci.click.echo") as echo:
+    with patch("click.echo") as echo:
         runtime.universal_config = Mock()
-        multi_cmd = cci.RunTaskCommand()
+        multi_cmd = task.RunTaskCommand()
         with click.Context(multi_cmd, obj=runtime) as ctx:
             multi_cmd.format_help(ctx, Mock())
         assert 4 == echo.call_count
@@ -118,10 +119,10 @@ def test_format_help(runtime):
 
 
 def test_get_default_command_options():
-    opts = RunTaskCommand()._get_default_command_options(is_salesforce_task=False)
+    opts = task.RunTaskCommand()._get_default_command_options(is_salesforce_task=False)
     assert len(opts) == 4
 
-    opts = RunTaskCommand()._get_default_command_options(is_salesforce_task=True)
+    opts = task.RunTaskCommand()._get_default_command_options(is_salesforce_task=True)
     assert len(opts) == 5
     assert any([o.name == "org" for o in opts])
 
@@ -130,7 +131,7 @@ def test_collect_task_options():
     new_options = {"debug-before": None}
     old_options = (("color", "green"),)
 
-    opts = RunTaskCommand()._collect_task_options(
+    opts = task.RunTaskCommand()._collect_task_options(
         new_options, old_options, "dummy-task", color_opts["options"]
     )
     assert opts == {"color": "green"}
@@ -141,7 +142,7 @@ def test_collect_task_options__duplicate():
     old_options = (("color", "green"),)
 
     with pytest.raises(CumulusCIUsageError):
-        RunTaskCommand()._collect_task_options(
+        task.RunTaskCommand()._collect_task_options(
             new_options, old_options, "dummy-task", color_opts["options"]
         )
 
@@ -151,9 +152,104 @@ def test_collect_task_options__not_in_task():
     old_options = (("color", "green"),)
 
     with pytest.raises(CumulusCIUsageError):
-        RunTaskCommand()._collect_task_options(
+        task.RunTaskCommand()._collect_task_options(
             new_options, old_options, "dummy-task", {"not-color": {}}
         )
+
+
+@patch("cumulusci.cli.task.CliTable")
+def test_task_list(cli_tbl):
+    runtime = Mock()
+    runtime.universal_config.cli__plain_output = None
+    runtime.get_available_tasks.return_value = [
+        {"name": "test_task", "description": "Test Task", "group": "Test Group"}
+    ]
+
+    run_click_command(task.task_list, runtime=runtime, plain=False, print_json=False)
+
+    cli_tbl.assert_called_with(
+        [["Task", "Description"], ["test_task", "Test Task"]],
+        "Test Group",
+        wrap_cols=["Description"],
+    )
+
+
+@patch("json.dumps")
+def test_task_list_json(json_):
+    task_dicts = {
+        "name": "test_task",
+        "description": "Test Task",
+        "group": "Test Group",
+    }
+    runtime = Mock()
+    runtime.universal_config.cli__plain_output = None
+    runtime.get_available_tasks.return_value = [task_dicts]
+
+    run_click_command(task.task_list, runtime=runtime, plain=False, print_json=True)
+
+    json_.assert_called_with([task_dicts])
+
+
+@patch("cumulusci.cli.task.doc_task", return_value="docs")
+def test_task_doc(doc_task):
+    runtime = Mock()
+    runtime.universal_config.tasks = {"test": {}}
+    run_click_command(task.task_doc, runtime=runtime, project=False)
+    doc_task.assert_called()
+
+
+def test_task_doc__project__outside_project():
+    runtime = Mock()
+    runtime.project_config = None
+    with pytest.raises(click.UsageError):
+        run_click_command(task.task_doc, runtime=runtime, project=True)
+
+
+@patch("click.echo")
+@patch("cumulusci.cli.task.doc_task", return_value="docs")
+def test_task_doc_project(doc_task, echo):
+    runtime = Mock()
+    runtime.universal_config = {"tasks": {}}
+    runtime.project_config = BaseProjectConfig(
+        runtime.universal_config,
+        {
+            "project": {"name": "Test"},
+            "tasks": {"task1": {"a": "b"}, "task2": {}},
+        },
+    )
+    runtime.project_config.config_project = {"tasks": {"task1": {"a": "b"}}}
+    run_click_command(task.task_doc, runtime=runtime, project=True)
+    doc_task.assert_called()
+    echo.assert_called()
+
+
+@patch("cumulusci.cli.task.Path")
+@patch("click.echo")
+@patch("cumulusci.cli.task.doc_task", return_value="docs")
+def test_task_doc_project_write(doc_task, echo, Path):
+    runtime = Mock()
+    runtime.universal_config.tasks = {"test": {}}
+    runtime.project_config = BaseProjectConfig(
+        runtime.universal_config,
+        {
+            "project": {"name": "Test"},
+            "tasks": {"option": {"a": "b"}},
+        },
+    )
+    runtime.project_config.config_project = {"tasks": {"option": {"a": "b"}}}
+    run_click_command(task.task_doc, runtime=runtime, project=True, write=True)
+    doc_task.assert_called()
+    echo.assert_not_called()
+
+
+@patch("cumulusci.cli.task.rst2ansi")
+@patch("cumulusci.cli.task.doc_task")
+def test_task_info(doc_task, rst2ansi):
+    runtime = Mock()
+    runtime.project_config.tasks__test = {"options": {}}
+    run_click_command(task.task_info, runtime=runtime, task_name="test")
+    doc_task.assert_called_once()
+    rst2ansi.assert_called_once()
 
 
 class SetTrace(Exception):
